Refactor using typestate pattern
