+++ src/lib.rs
@@ -1,3 +1,3 @@
 mod kibana_object_manager;
 
-pub use kibana_object_manager::{KibanaObjectManager, KibanaObjectManagerBuilder};
+pub use kibana_object_manager::KibanaObjectManagerBuilder;
